refactor initialState
