+++ src/reducers/books.js
@@ -1,11 +1,9 @@
 /* eslint-disable import/no-cycle */
-import { initialState } from './index';
 
-const books = (state = initialState, action) => {
+const books = (state, action) => {
   const newState = { ...state };
   switch (action.type) {
     case 'CREATE_BOOK':
-      //     console.log('Created Book!!');
       break;
     case 'DELETE_BOOK':
       break;
