+++ src/reducers/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 /* eslint-disable import/no-extraneous-dependencies */
 import { combineReducers } from 'redux';
 import books from './books';
